Add features from test-parsec.sh

+++ test-graphchi.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
 if [ $UID -ne 0 ]; then
-	echo "This script must be run as root."
-	exit 1
+    echo "This script should be run as root."
+    sleep 1
 fi
 
 # get_runtime <param>
@@ -21,23 +21,57 @@ function stdev() {
 	echo ${@:1} | tr " " "\\n" | awk '{sum+=$1; sumsq+=$1*$1}END{print sqrt(sumsq/NR - (sum/NR)**2)}' 
 }
 
+# cleanup $dir
+function cleanup() {
+        local cdir=$1
+	if [[ $cdir = /u/$(whoami)/* ]] || [[ $(pwd) = /localhost/$(whoami)/* ]]; then
+		find $cdir -type f -user root -exec rm -f {} \;
+		find $cdir -type d -user root -exec rm -rf {} \;
+	else
+		echo "Refusing to destroy contents of non-user directory $cdir"
+	fi
+}
+
 # run_test <appname> <runtime-param> <cmd>
 function run_test() {
 	local appname=$1
-	local runtime_param=$2
+        local runtime_param=$2
 	local cmd=${@:3}
 	local stats=`readlink -f stats/${appname}-runtimes.txt`
 	local perf_stats=`readlink -f stats/${appname}-rhm.txt`
 	local task_mapper=`readlink -f ./Task_mapper2`
-	local colocated_sched=`readlink -f colocated.sched`
-	local spread_sched=`readlink -f spread.sched`
+	local colocated_sched='default:colocated' #`readlink -f colocated.sched`
+	local spread_sched='default:spread' #`readlink -f spread.sched`
 	local count=4
+
+        if [ ! -e $task_mapper ]; then
+            make CFLAGS=-std=gnu99 LDFLAGS=-lm Task_mapper2
+        fi
+
+        if [ ! -e $(dirname $stats) ]; then
+            mkdir -p $(dirname $stats)
+        fi
+
+        #if [ ! -e $colocated_sched ]; then
+        #    echo "$colocated_sched does not exist!"
+        #    exit 1
+        #fi
+
+        #if [ ! -e $spread_sched ]; then
+        #    echo "$spread_sched does not exist!"
+        #    exit 1
+        #fi
 
 	cd graphchi-cpp
 
-	chown $SUDO_USER $stats
-	chown $SUDO_USER $perf_stats
-	cat <(echo $appname) <(perl -e "printf '-' x ($(wc -m <<< $appname) - 1)") <(echo "") | tee $stats $perf_stats 1>/dev/null
+        export GRAPHCHI_DIR=$(pwd)
+	trap "{ cleanup $GRAPHCHI_DIR; rm -f ${appname}-pipe; pkill -u root,pferro --signal TERM Task_mapper2; exit 0; }" EXIT SIGINT SIGTERM
+
+        if [ $UID -eq 0 ]; then
+            chown $SUDO_USER $stats
+            chown $SUDO_USER $perf_stats
+        fi
+	cat <(echo $appname) <(perl -e "printf '-' x ($(wc -m <<< $appname) - 1)") <(echo "") <(echo "Command: $cmd") <(echo "") | tee $stats $perf_stats 1>/dev/null
 
 	schednames=('Colocated' 'Spread')
 	schedules=($colocated_sched $spread_sched)
@@ -51,7 +85,7 @@ function run_test() {
 		local runtimes=()
 
 		for i in $(seq 1 $count); do
-			echo edgelist | $cmd
+                        echo edgelist | $cmd
 			tm=$(get_runtime $runtime_param)
 			runtimes+=($tm)
 			echo "$tm s" >> $stats
@@ -60,12 +94,16 @@ function run_test() {
 		echo "" >> $stats
 		echo "`avg ${runtimes[@]}` s (avg) +/- `stdev ${runtimes[@]}`" >> $stats
 		echo "" >> $stats
-
-		# count REMOTE_HIT_MODIFIED (r10d3) hardware counter
-		mkfifo ${appname}-pipe
-		cat ${appname}-pipe >> $perf_stats &
-		echo edgelist | perf stat -e r10d3 -a --per-core -o ${appname}-pipe $cmd
-		rm ${appname}-pipe
+                
+                if [ $UID -eq 0 ]; then
+                    # count REMOTE_HIT_MODIFIED (r10d3) hardware counter
+                    mkfifo ${appname}-pipe
+                    cat ${appname}-pipe >> $perf_stats &
+                    echo edgelist | perf stat -e r10d3 -a --per-core -o ${appname}-pipe $cmd 1>/dev/null
+                    rm ${appname}-pipe
+                else
+                    echo "Warning: Skipping REMOTE_HIT_MODIFIED test because we lack permissions to run perf-stat"
+                fi
 
 		kill -s TERM $child_pid
 		echo "Waiting for process $child_pid to terminate..."
@@ -74,6 +112,8 @@ function run_test() {
 	done
 
 	cd ..
+
+	exit 0
 }
 
 run_test pagerank runtime bin/example_apps/pagerank file /u/pferro/Downloads/soc-LiveJournal1.txt niters 10
